Changed fragment names to their respective tabs
